+++ packages/verified-fetch/src/utils/libp2p-defaults.browser.ts
@@ -2,13 +2,14 @@ import { libp2pDefaults } from '@helia/libp2p'
 import { webRTCDirect } from '@libp2p/webrtc'
 import { webSockets } from '@libp2p/websockets'
 import type { ServiceFactoryMap } from './libp2p-types.ts'
-import type { DefaultLibp2pServices } from '@helia/libp2p'
+import type { CreateLibp2pOptions, DefaultLibp2pServices } from '@helia/libp2p'
 import type { Libp2pOptions } from 'libp2p'
 
 type ServiceMap = Pick<DefaultLibp2pServices, 'dcutr' | 'identify' | 'identifyPush' | 'keychain' | 'ping'>
 
-export function getLibp2pConfig (): Libp2pOptions & Required<Pick<Libp2pOptions, 'services'>> {
-  const libp2pDefaultOptions = libp2pDefaults()
+export function getLibp2pConfig (options?: CreateLibp2pOptions): Libp2pOptions & Required<Pick<Libp2pOptions, 'services'>> {
+  // @ts-expect-error cannot derive correct type
+  const libp2pDefaultOptions = libp2pDefaults(options)
 
   libp2pDefaultOptions.start = false
   libp2pDefaultOptions.addresses = { listen: [] }
