+++ packages/verified-fetch/src/utils/libp2p-defaults.ts
@@ -1,20 +1,17 @@
 import { libp2pDefaults } from '@helia/libp2p'
-import { kadDHT } from '@libp2p/kad-dht'
 import type { ServiceFactoryMap } from './libp2p-types.ts'
-import type { DefaultLibp2pServices } from '@helia/libp2p'
+import type { CreateLibp2pOptions, DefaultLibp2pServices } from '@helia/libp2p'
 import type { Libp2pOptions } from 'libp2p'
 
-type ServiceMap = Pick<DefaultLibp2pServices, 'autoNAT' | 'dcutr' | 'dht' | 'identify' | 'identifyPush' | 'keychain' | 'ping' | 'upnp'>
+type ServiceMap = Pick<DefaultLibp2pServices, 'autoNAT' | 'dcutr' | 'identify' | 'identifyPush' | 'keychain' | 'ping' | 'upnp'>
 
-export function getLibp2pConfig (): Libp2pOptions & Required<Pick<Libp2pOptions, 'services'>> {
-  const libp2pDefaultOptions = libp2pDefaults()
+export function getLibp2pConfig (options?: CreateLibp2pOptions): Libp2pOptions & Required<Pick<Libp2pOptions, 'services'>> {
+  // @ts-expect-error cannot derive correct type
+  const libp2pDefaultOptions = libp2pDefaults(options)
 
   const services: ServiceFactoryMap<ServiceMap> = {
     autoNAT: libp2pDefaultOptions.services.autoNAT,
     dcutr: libp2pDefaultOptions.services.dcutr,
-    dht: kadDHT({
-      clientMode: true
-    }),
     identify: libp2pDefaultOptions.services.identify,
     identifyPush: libp2pDefaultOptions.services.identifyPush,
     keychain: libp2pDefaultOptions.services.keychain,
